+++ app/(root)/workspace/meeting/page.tsx
@@ -1,44 +1,7 @@
-'use client';
-
-import { useEffect, useState } from 'react';
-
-import { getMeetingRoomList } from '@/apis/meeting';
-import { getWorkspaceInfo } from '@/apis/workspace';
 import CreateMeetingButton from '@/components/meeting/waitingRoom/CreateMeetingButton';
-import EmptyWaitingRoomTemplate from '@/components/meeting/waitingRoom/EmptyWaitingRoomTemplate';
-import ListWaitingRoomTemplate from '@/components/meeting/waitingRoom/ListWaitingRoomTemplate';
-import { useWebsocket } from '@/lib/websocket/WebsocketProvider';
-import { RoomList } from '@/models/meeting/entity/meeting';
+import MeetingRoomList from '@/components/meeting/waitingRoom/MeetingRoomList';
 
 const MeetingPage = () => {
-	const [roomList, setRoomList] = useState<RoomList>();
-	const stompClient = useWebsocket();
-
-	useEffect(() => {
-		const handleConnect = async () => {
-			const { workspaceId } = await getWorkspaceInfo();
-			subscribeMeetingList(workspaceId);
-			getMeetingRoomList();
-		};
-
-		handleConnect();
-	}, []);
-
-	const subscribeMeetingList = (workspaceId: string) => {
-		stompClient.subscribe(`/topic/${workspaceId}/meetingRoomList`, function (message) {
-			const body = JSON.parse(message.body);
-			setRoomList(body.meetingRoomList);
-		});
-	};
-
-	const renderMeetingRoomList = () => {
-		if (!roomList || roomList.length === 0) {
-			return <EmptyWaitingRoomTemplate />;
-		}
-
-		return <ListWaitingRoomTemplate roomList={roomList} />;
-	};
-
 	return (
 		<div className="w-full">
 			<div>
@@ -50,7 +13,7 @@ const MeetingPage = () => {
 				<h6 className="typo-SubHeader1">진행중인 회의</h6>
 				<CreateMeetingButton />
 			</div>
-			{renderMeetingRoomList()}
+			<MeetingRoomList />
 		</div>
 	);
 };
